+++ DataGenerator/Data/DataInfo.cs
@@ -45,7 +45,7 @@ namespace DataGenerator.Data
             universalEvent.ElementEffects.Add(-10, new PreTypeEvent());
             universalEvent.OnDeaths.Add(-10, new PreDeathEvent());
             universalEvent.OnDeaths.Add(-10, new SetDeathEvent());
-            universalEvent.OnDeaths.Add(0, new ImpostorReviveEvent("imposter"));
+            universalEvent.OnDeaths.Add(0, new ImpostorReviveEvent("imposter", "transformed"));
             universalEvent.OnDeaths.Add(10, new HandoutRelativeExpEvent(1, 7, 10, 5, 2));
             universalEvent.OnMapStarts.Add(-10, new FadeInEvent());
             universalEvent.OnMapStarts.Add(-5, new SpecialIntroEvent());