+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.7.14 Changes #
 
 * Secret Garden exit moved one floor earlier
+* Prevent Transform's effect when already transformed
+* Imposter's revive is now 1/2 the HP before transformation
 * Dev: Made it possible to apply universal zone steps
 * Dev: Made it possible to add friendly NPCs in static maps
 * Dev: Issue fixed where editing TypeRecruitmentEvent was impossible